If the clock changes substantially then the calculated time between "now" and the last IAX2 registration or the last DTE event or the last RING will be very "big" and weird.
Apparently this does not pose a problem for most systems.  However, there are some environments where a select() with one of these "big" values as the timeval will lead to an endless loop and 100% CPU consumption because the select() will never block and the "last" event time does not get updated.

This fixes that.