ITU-T V.250 6.3.1 defines the ATD command.  It states:
"Any characters appearing in the dial string that the DCE does not recognize as a valid part of the call addressing information or as a valid modifier shall be ignored. This permits characters such as parentheses and hyphens to be included that are typically used in formatting of telephone numbers."

The permissible dialing digits are defined in the following section:
6.3.1.1.  The "+" character is permissible, as with *, #, A, B, C, and D.  However, spandsp is removing/ignoring the "+".

This should get it to pass the "+" along.