These changes are a contribution from Robert Boisvert and the Alpine Linux team to address issues that were encountered while running iaxmodem in valgrind.
See: https://sourceforge.net/p/iaxmodem/discussion/497500/thread/b2469a0d1a/

The initialization matter for tioflags is self-evident.

The change in libiax2 does not exactly match a change to the same code in the upstream, but it seems likely to accomplish the same.

The change in spandsp was stated to have come from upstream ("a later version of this library"); however, the latest version of the library from soft-switch.org (before it went off-line) did not contain this change.  So, the comment must refer to another source.  However, the need for a fix is clear, and so we'll go with this as it is known to address the reporter's problem.

In case it needs to be said here, libiax2 upstream development stopped many, many years ago.  The last release of the upstream library was tested with iaxmodem a long time ago and found to be inferior or problematic, and so the upstream libiax2 was not followed.

As for spandsp, there was generally a fairly close development between iaxmodem's copy and the upstream spandsp source, but Steve Underwood has been unresponsive for years, and his websites have begun to go off-line now.  To the best of my knowledge iaxmodem uses the latest version of Steve's work in spandsp.  However, as time goes on it will become necessary to continue without his guidance, and a new upstream code repository is needed.