Take 3.  This correctly addresses the issue that was attempted to be fixed by r50 and r51.
If a hangup occurred after the DTE sent DLE+ETX (following AT+FTM=n + CONNECT, for example) but before all of the data had been transmitted and the OK result given, then that would trigger a problem.  This was most common in sending TCF because it was so short.

For some reason unknown to me t31.c had the modem being put in command mode before the OK message... this resulted in the lack of the NO CARRIER message on hangup.

Removing the problematic command mode instigator seems to be harmless as it seems to have been superfluous - it was being put in command mode again with the OK.